+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##  2019-05-28  
+
+New provide eos-transit-portis-provider
+
+Updated scatterjs-core@2.7.16 to scatterjs-core@2.7.18 
 
 ## [3.3.2 - eos-transit-ledger-provider] - 2019-05-28